Point distance calculation

+++ Searcher/src/main/java/com/github/pathfinder/data/Coordinate.java
@@ -1,6 +1,5 @@
 package com.github.pathfinder.data;
 
-public record Coordinate(Integer longitude,
-                         Integer latitude) {
+public record Coordinate(Double longitude, Double latitude) {
 
 }
