+++ Searcher/src/main/java/com/github/pathfinder/data/point/Point.java
@@ -2,9 +2,9 @@ package com.github.pathfinder.data.point;
 
 import com.github.pathfinder.database.entity.LandType;
 
-public record Point(Integer altitude,
-                    Integer longitude,
-                    Integer latitude,
+public record Point(Double altitude,
+                    Double longitude,
+                    Double latitude,
                     LandType landType) {
 
 }
